move cert temp file logic to cluster crate
