+++ crates/fluvio-cluster/Cargo.toml
@@ -53,6 +53,7 @@ anyhow = "1.0.44"
 async-channel = "1.6.1"
 indicatif = "0.16.2"
 nix = "0.23"
+rand = "0.8.4"
 
 # Fluvio dependencies
 fluvio = { version = "0.10.0", path = "../fluvio", default-features = false }
