+++ crates/fluvio/Cargo.toml
@@ -41,7 +41,6 @@ siphasher = "0.3.5"
 cfg-if = "1.0.0"
 derive_builder = "0.10"
 async-trait = "0.1.51"
-rand = "0.8.4"
 
 # Fluvio dependencies
 fluvio-future = { version = "0.3.12", features = ["task", "openssl_tls", "task_unstable"] }
